requirements.txt
click==8.0.3
numpy==1.21.3
pandas==1.3.4
psycopg2-binary==2.9.1
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
six==1.16.0
